Extracted the GetTypeSelection method
